Reorganized the file structure and placed all components into the LandingPage component
