+++ src/Components/Body/LandingPage.js
@@ -2,7 +2,8 @@
 
 import React, { Component } from "react";
 import SunriseImage from "../../Images/Sunrise.jpeg";
-import Navbar from "../NavBar/NavBar";
+import Navbar from "./NavBar";
+import Footer from "./Footer";
 import styled from "@emotion/styled";
 
 const LandingPagePicture = styled.div`
@@ -38,6 +39,7 @@ class LandingPage extends Component {
             culpa qui officia deserunt mollit anim id est laborum."
           </MissionText>
         </LandingPagePicture>
+        <Footer />
       </>
     );
   }
